Move createClusterNode and slot functions to cluster_state.c
Move createClusterNode, clusterNodeSetSlotBit, clusterAddSlot,
clusterDelSlot, and clusterDelNodeSlots from cluster_legacy.c to
cluster_state.c. These are protocol-agnostic state management functions.

createClusterNode uses the initNodeData vtable callback to let the
protocol implementation attach its data.

The legacy gossip owner_not_claiming_slot bitmap clearing, previously
done inside clusterAddSlot/clusterDelSlot, is now done explicitly at
each call site in cluster_legacy.c.

Remove the onSlotReset vtable callback which is no longer needed.
